Fully refactored info fetch, now kinda working
