+++ resources/locales/de.json
@@ -21,7 +21,7 @@
     "t_english": "Englisch",
     "t_spanish": "Spanisch",
     "t_german": "Deutsche",
-    "t_footer": "PokéInfo beta v13",
+    "t_footer": "PokéInfo beta v14",
     "t_langText": "Sprache:",
     "t_langCode": "DE",
     "t_anyType": "-- JEDER --",
