+++ resources/locales/en.json
@@ -21,7 +21,7 @@
     "t_english": "English",
     "t_spanish": "Spanish",
     "t_german": "German",
-    "t_footer": "PokéInfo beta v13",
+    "t_footer": "PokéInfo beta v14",
     "t_langText": "Language:",
     "t_langCode": "EN",
     "t_anyType": "-- ANY --",
