+++ resources/locales/es.json
@@ -21,7 +21,7 @@
     "t_english": "Inglés",
     "t_spanish": "Español",
     "t_german": "Alemán",
-    "t_footer": "PokéInfo beta v13",
+    "t_footer": "PokéInfo beta v14",
     "t_langText": "Idioma:",
     "t_langCode": "ES",
     "t_anyType": "-- CUALQUIERA --",
